// leoz-web/src/app/core/models/touroptimizing.model.ts
export interface Tour {
  deliverylistNumber?: number;
  shipments?: number;
  packages?: number;
  weight?: number;
  time?: string;
  distance?: string;
  optimized?: boolean;
  zipCode?: string;
  timeRange?: string;
  tourNo?: number;
  tourWeight?: number;
  tourShipments?: number;
  tourPackages?: number;
  tourTime?: string;
  tourDistance?: string;
  tourOptimized?: boolean;
}